Switch cursor time from int64_t to double

+++ src/cursor.h
@@ -10,7 +10,7 @@
 typedef struct _iproc_cursor iproc_cursor;
 
 struct _iproc_cursor {
-    int64_t             tcur;
+    double              tcur;
     iproc_history      *history;
     iproc_message_iter *it;
     iproc_refcount      refcount;
@@ -26,7 +26,7 @@ int             iproc_cursor_next       (iproc_cursor   *cursor);
 int             iproc_cursor_started    (iproc_cursor   *cursor);
 int             iproc_cursor_finished   (iproc_cursor   *cursor);
 
-int64_t         iproc_cursor_time       (iproc_cursor   *cursor);
+double          iproc_cursor_time       (iproc_cursor   *cursor);
 iproc_history * iproc_cursor_history    (iproc_cursor   *cursor);
 int64_t         iproc_cursor_nmsg       (iproc_cursor   *cursor);
